requirements.txt
absl_py==2.1.0
apache_beam==2.54.0
einops==0.7.0
imageio==2.34.0
ipython==8.12.3
krippendorff==0.6.1
librosa==0.10.1
matplotlib==3.8.3
nltk==3.8.1
numpy==1.24.4
openai_clip==1.0.1
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.2.1
Pillow==10.2.0
plotly==5.20.0
pytorch_lightning==2.2.1
resampy==0.4.3
scikit_learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soundfile==0.12.1
statsmodels==0.14.1
tensorflow==2.16.1
tensorflow_hub==0.16.1
tf_slim==1.1.0
torch==2.2.1
torchaudio==2.2.1
torchmetrics==1.3.1
torchnet==0.0.4
torchvision==0.17.1
tqdm==4.66.2
